+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 
 
 ## 2.1.11.1-SNAPSHOT (upcoming)
+
  * Fixed bitemporal bug (issue#46)
  * Fixed default directory path
-  
+ * Added query builder module (issue#50)
+ 
 ## 2.1.11.0 (27 October 2015)
 
  * Upgrade to Apache Cassandra 2.1.11